feature: add product services repository

+++ src/main/java/com/platzi/market/persistence/entity/Cliente.java
@@ -18,7 +18,7 @@ public class Cliente {
 
   private String apellidos;
 
-  private Integer celular;
+  private Long celular;
 
   private String direccion;
 
@@ -52,11 +52,11 @@ public class Cliente {
     this.apellidos = apellidos;
   }
 
-  public Integer getCelular() {
+  public Long getCelular() {
     return celular;
   }
 
-  public void setCelular(Integer celular) {
+  public void setCelular(Long celular) {
     this.celular = celular;
   }
